backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
google-genai>=1.5.0
python-multipart==0.0.20
websockets==14.1
